+++ aioplus/internal/utils/enums.py
@@ -4,4 +4,5 @@ from enum import Enum, auto
 class AioPlus(Enum):
     """Sentinel values."""
 
+    EMPTY = auto()
     SENTINEL = auto()