+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Added `aioplus.atriplewise`;
+* Added `aioplus.awindowed`.
+
 ## [0.1.0] - 2025-08-09
 
 * Added `aioplus.acycle`;